Addenda to UNIX 7th edition distribution tape, 7/27/79.
Format: tar(1), 800 bpi.

Contents:

README: this descriptive file.

lp.c: the missing line printer driver that
      belongs in /usr/sys/dev/lp.c.
      Notice the comment at head of program.

lpr: a directory containing the lpr utility and
      its daemon lpd.
      See lpr/makefile for instructions on putting it
      together.

lpd.8: the manual section for the line printer daemon.

c2fix: instructions for repairing a bug in the
      C optimizer.

Other remarks:

The directory /usr/23 mentioned in the schedule
of files with the license agreement is actually
known as /usr/40.

The following bugs exist, no fix is included.
      (1) adb does not report floating registers correctly
      (2) ldiv, lmod fail with largest negative dividend
          (these implement division of longs in C)
      (3) dump(1) maintains ddate incorrectly.
          This bug is relatvely innocuous; it causes
          more dumping than necessary on some occasions.

Signed-off-by: Michael van der Westhuizen <rmikey@fb.com>