Add Raspberry Pi 1 kernel port files
This is not quite the whole picture (more kernel mods follow),
but this commit includes the lion's share of the kernel port.

bcm283x/rt is from the LLVM compiler-rt project, and the
sd card driver is a heavily modified version of the LdB
driver code (with FAT bits stripped out).

Most of what's here still needs significant cleanup.

Signed-off-by: Michael van der Westhuizen <rmikey@fb.com>