Add VFP management
Hook up the undefined handler, using it to turn on the VFP on-demand.

Unlike the approach taken in modern operating systems, and due to the fp
regs being stored in swappable memory, save the fp regs on context switch,
ensuring that any changes are persisted.  When restoring regs, just turn
the VFP off, allowing the undefined instruction handler to turn VFP on
in an on-demand way.

All of this is to say, only quanta that use the VFP unit will be subject
to VFP overhead on context switch, but not all modern lazy FP switching
techniques are possible in this version of UNIX.

The basic approach is heavily inspired by the BSDs, with some non-trival
rework to get things working in this codebase with this approach.

Signed-off-by: Michael van der Westhuizen <r1mikey@gmail.com>