emmc: add an alternate SD card driver implementation
Add an SD card driver that separates the generic SD/MMC host driver
from the BCM2835-specific code.  This new driver also correctly deals
with clocks on the device and addresses a handful of other shortcomings
in the previous code.

Disable the older code for now, with an eye to removing it soon.

Signed-off-by: Michael van der Westhuizen <r1mikey@gmail.com>