Optimise printf/putchar kernel output
The putchar function presently saves and restores a lot of state, but
not in a way that agrees with the TRM view of how a UART should be
treated, and in a loop (via the functions in prf.c).

Create state helpers to save and restire the interrupt mask, as all
other registers have already been set to their desired values.

Use this new helper in prf.c, outside the printf loop, so that
state is only saved and restored once per kernel printf.

While we're here, start macro-ing up register bit definitions to
make this code more understandable.

Signed-off-by: Michael van der Westhuizen <r1mikey@gmail.com>